Added utils.js that is responsible for common exported js methods
