+++ frontend/themes/light_theme/components/javascript/design.js
@@ -1,0 +1,4 @@
+// DESIGN IDEAS FOR CHARTS
+// https://echarts.apache.org/examples/en/editor.html?c=pie-nest - Black border
+
+// Works fine, require some manipulations with label/tooltip formatter
