--- frontend/themes/light_theme/components/javascript/test.js
@@ -1,5 +1,0 @@
-import { greet } from './fillChart.js';
-
-window.callMe = function callMe() {
-    greet('John');
-}
