+++ src/main/dev-bundle/config/stats.json
@@ -216,12 +216,11 @@
   "Frontend/generated/jar-resources/vaadin-grid-flow-selection-column.js",
   "Frontend/generated/jar-resources/vaadin-time-picker/timepickerConnector.js",
   "Frontend/generated/jar-resources/virtualListConnector.js",
-  "Frontend/themes/light_theme/components/javascript/fillChart.js",
-  "Frontend/themes/light_theme/components/javascript/test.js"
+  "Frontend/themes/light_theme/components/javascript/fillPieChart.js"
  ],
  "frontendHashes": {
-  "themes/light_theme/components/javascript/fillChart.js": "4e111be7f20cb81596191ee6d964834a18093ad426ecdb9f67854dce2e293582",
-  "themes/light_theme/components/javascript/test.js": "bcee8fe9f7f853244951e393ba745ad1eb40d64f635d3ab1d4e13cc1ffd5f42a",
+  "themes/light_theme/components/javascript/utils.js": "1129730be2854be705af7ee57fa497c5fbc89a974689876ac40599bec1094d22",
+  "themes/light_theme/components/javascript/fillPieChart.js": "38c9a9fa1ae3661d35da2078d4d681289debca774b77c23e653b5a0c125a4e64",
   "comboBoxConnector.js": "1d004c3dae9c76f8bcc68678abe1b8f7bdc8b456f5c9c4c7f190df8ed25d66cd",
   "contextMenuConnector.js": "25a200f8086461635ddd92316ebeb5a868294c8650a190efbae120edb1088f4f",
   "contextMenuTargetConnector.js": "870f67643988dd5618c247bd3c28d843f70aa026ee89f44c3c7556b5e70856af",
@@ -244,7 +243,7 @@
   "light_theme": "{\n  \"lumoImports\" : [ \"typography\", \"color\", \"spacing\", \"badge\", \"utility\" ]\n}"
  },
  "entryScripts": [
-  "VAADIN/build/indexhtml-56d16e7e.js"
+  "VAADIN/build/indexhtml-5414e54e.js"
  ],
  "webComponents": [],
  "packageJsonHash": "a89b689ec344ccdc7b126a3605399c2171141918d2911a0efd1d21aa157c62b8"
