+++ src/main/dev-bundle/webapp/index.html
@@ -3,6 +3,9 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap" rel="stylesheet">
   <style>
     body, #outlet {
       height: 100vh;
@@ -10,16 +13,16 @@
       margin: 0;
     }
   </style>
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap" rel="stylesheet">
   
-  <script type="module" crossorigin src="./VAADIN/build/indexhtml-56d16e7e.js"></script>
+  <script type="module" crossorigin src="./VAADIN/build/indexhtml-5414e54e.js"></script>
 </head>
+
 <body>
-  <!-- This outlet div is where the views are rendered -->
-  <div id="outlet"></div>
+  <div id="outlet">
+    <!-- This outlet div is where the views are rendered -->
+  </div>
 </body>
+
 </html>
